+++ libraries/objects/src/checkIsPlainObject.ts
@@ -2,8 +2,13 @@
  * Returns true if the given value is a plain object (i.e., an object created with the `Object`
  * constructor or with the object literal syntax, and not a function, class instance, null, or array).
  */
-export function isPlainObject(value: PlainObject | unknown): value is PlainObject {
+export function checkIsPlainObject<T extends PlainObject>(value: T | unknown): value is T {
   return value instanceof Object && Object.getPrototypeOf(value) === Object.prototype;
 }
+
+/**
+ * @deprecated Use `checkIsPlainObject` instead.
+ */
+export const isPlainObject = checkIsPlainObject;
 
 export type PlainObject = { [key: string]: unknown } & ({ bind?: never } | { call?: never });